+++ maat-court-data-api/src/main/java/gov/uk/courtdata/hearing/impl/HearingResultedImpl.java
@@ -12,7 +12,7 @@ import gov.uk.courtdata.model.Result;
 import gov.uk.courtdata.model.hearing.HearingResulted;
 import gov.uk.courtdata.processor.OffenceCodeRefDataProcessor;
 import gov.uk.courtdata.processor.ResultCodeRefDataProcessor;
-import gov.uk.courtdata.prosecutionconcluded.helper.OffenceHelper;
+import gov.uk.courtdata.offence.helper.OffenceHelper;
 import gov.uk.courtdata.repository.IdentifierRepository;
 import gov.uk.courtdata.repository.WqLinkRegisterRepository;
 import lombok.RequiredArgsConstructor;
