+++ maat-court-data-api/src/test/java/gov/uk/courtdata/hearing/impl/HearingResultedImplTest.java
@@ -11,9 +11,9 @@ import gov.uk.courtdata.model.Defendant;
 import gov.uk.courtdata.model.Offence;
 import gov.uk.courtdata.model.Result;
 import gov.uk.courtdata.model.hearing.HearingResulted;
+import gov.uk.courtdata.offence.helper.OffenceHelper;
 import gov.uk.courtdata.processor.OffenceCodeRefDataProcessor;
 import gov.uk.courtdata.processor.ResultCodeRefDataProcessor;
-import gov.uk.courtdata.prosecutionconcluded.helper.OffenceHelper;
 import gov.uk.courtdata.repository.IdentifierRepository;
 import gov.uk.courtdata.repository.WqLinkRegisterRepository;
 import org.junit.jupiter.api.BeforeEach;
