+++ maat-court-data-api/src/test/java/gov/uk/courtdata/helper/ReservationsRepositoryHelperTest.java
@@ -1,4 +1,4 @@
-package gov.uk.courtdata.prosecutionconcluded.helper;
+package gov.uk.courtdata.helper;
 
 import gov.uk.courtdata.entity.ReservationsEntity;
 import gov.uk.courtdata.repository.ReservationsRepository;
